Project_GP1/requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
langchain
chroma